fonc nombrepersonne

+++ src/OffreBundle/Controller/OffreController.php
@@ -1,7 +1,8 @@
 <?php
 
 namespace OffreBundle\Controller;
-
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
+use FOS\UserBundle\Model\UserInterface;
 use OffreBundle\Entity\Offre;
 use OffreBundle\Form\OffreType;
 use ProjetBundle\Entity\Projet;
@@ -12,6 +13,10 @@ class OffreController extends Controller
 {
     public function afficheAction()
     {
+        $user = $this->getUser();
+        if (!is_object($user) || !$user instanceof UserInterface) {
+            throw new AccessDeniedException('This user does not have access to this section.');
+        }
         $em = $this->getDoctrine()->getManager();
 
         $projets = $em->getRepository('ProjetBundle:Projet')->findAll();
@@ -23,6 +28,10 @@ class OffreController extends Controller
 
     public function showAction(Projet $projet)
     {
+        $user = $this->getUser();
+        if (!is_object($user) || !$user instanceof UserInterface) {
+            throw new AccessDeniedException('This user does not have access to this section.');
+        }
 
         return $this->render('@Offre/offre/show.html.twig', array(
             'projet' => $projet,
@@ -31,6 +40,10 @@ class OffreController extends Controller
 
     public function mesoffresAction()
     {
+        $user = $this->getUser();
+        if (!is_object($user) || !$user instanceof UserInterface) {
+            throw new AccessDeniedException('This user does not have access to this section.');
+        }
         $em = $this->getDoctrine()->getManager();
 
         $offres = $em->getRepository('OffreBundle:Offre')->findAll();
@@ -42,11 +55,16 @@ class OffreController extends Controller
 
     public function postulerAction(Request $request, Projet $projet)
     {
+        $user = $this->getUser();
+        if (!is_object($user) || !$user instanceof UserInterface) {
+            throw new AccessDeniedException('This user does not have access to this section.');
+        }
         $offre = new Offre();
         $form = $this->createForm(OffreType::class,$offre);
         $form = $form->handleRequest($request);
         if ($form->isValid()) {
             $offre->setIdprojet($projet);
+            $offre->setIdfreelancer($user);
             $em = $this->getDoctrine()->getManager();
             $em->persist($offre);
             $em->flush();
@@ -60,6 +78,10 @@ class OffreController extends Controller
 
     public function deleteAction($id)
     {
+        $user = $this->getUser();
+        if (!is_object($user) || !$user instanceof UserInterface) {
+            throw new AccessDeniedException('This user does not have access to this section.');
+        }
         $em=$this->getDoctrine()->getManager();
         $offre=$em->getRepository(Offre::class)->find($id);
         $em->remove($offre);
@@ -69,6 +91,10 @@ class OffreController extends Controller
 
     public function updateAction(Request $request,$id)
     {
+        $user = $this->getUser();
+        if (!is_object($user) || !$user instanceof UserInterface) {
+            throw new AccessDeniedException('This user does not have access to this section.');
+        }
         $em=$this->getDoctrine()->getManager();
         $offre=$em->getRepository(Offre::class)->find($id);
         $form=$this->createForm(OffreType::class,$offre);
